refactor ReplyCommand
This moves the static methods clear_my_address and ensure_unique_address
out of ReplyCommand and into alot.db.utils.
